tests/wpt/tests/html/canvas/offscreen/fill-and-stroke-styles/2d.gradient.interpolate.multiple.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.gradient.interpolate.multiple</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.gradient.interpolate.multiple</h1>
<p class="desc"></p>


<script>
var t = async_test("");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  canvas.width = 200;
  var g = ctx.createLinearGradient(0, 0, 200, 0);
  g.addColorStop(0, '#ff0');
  g.addColorStop(0.5, '#0ff');
  g.addColorStop(1, '#f0f');
  ctx.fillStyle = g;
  ctx.fillRect(0, 0, 200, 50);
  _assertPixelApprox(canvas, 50,25, 127,255,127,255, 3);
  _assertPixelApprox(canvas, 100,25, 0,255,255,255, 3);
  _assertPixelApprox(canvas, 150,25, 127,127,255,255, 3);
  t.done();

});
</script>
